First AI

+++ BurgerTime/LadderComp.cpp
@@ -31,7 +31,7 @@ void LadderComp::RecalculateSize()
 	auto top = std::make_shared<dae::GameObject>();
 
 	top->SetParent(m_pGameObject);
-	top->SetPosition(0, m_NrOfSteps * m_StepSize.y);
+	top->SetPosition(0, (float)(m_NrOfSteps * m_StepSize.y));
 	top->SetSize(m_StepSize.x, m_StepSize.y);
 
 #if _DEBUGRENDERING
